requirements.txt
altair==5.0.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
comm==0.1.3
contourpy==1.1.0
cvxopt==1.3.1
cvxpy==1.3.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
ecos==2.0.12
executing==1.2.0
fancyimpute==0.7.0
fastjsonschema==2.16.3
fonttools==4.41.1
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
impyute==0.0.8
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keras==2.12.0
kiwisolver==1.4.4
knnimpute==0.1.0
lightgbm==4.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
mistune==2.0.5
mtcnn==0.1.1
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
nose==1.3.7
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.25.1
opencv-python==4.7.0.72
osqp==0.6.3
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.4.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
qdldl==0.1.7.post0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
scikit-learn==1.3.0
scipy==1.11.1
scs==3.2.3
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
streamlit==1.25.0
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.2.0
urllib3==2.0.3
validators==0.20.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
xgboost==1.7.6
zipp==3.16.2
